Pass precise position timestamp within `PlayerEvent`
The following `PlayerEvent::PositionChanged/SeekDone` events
should contains the precise position timestamps to pass them
to back to client, otherwise there are no possibility to
identify timestamp changes between two `PositionChanged` events
on the client side.

gst (5.104499333) -> servo (5)
gst (5.153333333) -> servo (5)

+++ examples/play_media_stream.rs
@@ -62,7 +62,7 @@ fn run_example(servo_media: Arc<ServoMedia>) {
             },
             PlayerEvent::VideoFrameUpdated => eprint!("."),
             PlayerEvent::PositionChanged(p) => {
-                if p == 4 {
+                if p as u64 == 4 {
                     break;
                 }
                 println!("Position changed {:?}", p)
